requirements.txt
Flask-SQLAlchemy
flask-restful
flask-googlemaps
xlsxwriter